requirements.txt
PyInquirer
pyyaml